split FE integration tests to per field test
