+++ x-pack/solutions/security/plugins/security_solution/public/detection_engine/rule_management_ui/pages/rule_management/__integration_tests__/rules_upgrade/upgrade_rule_after_preview/rule_name_override.test.ts
@@ -1,0 +1,61 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { mockAvailableDataViews } from '../mock/rule_upgrade_flyout';
+import { assertRuleUpgradePreview } from '../mock/assert_rule_upgrade_preview';
+import { assertRuleUpgradeAfterReview } from '../mock/assert_rule_upgrade_after_review';
+
+describe('Upgrade rule after preview - "rule_name_override" field', () => {
+  beforeAll(() => {
+    mockAvailableDataViews([], {
+      resolvedString: {
+        name: 'resolvedStringField',
+        type: 'string',
+        searchable: true,
+        aggregatable: true,
+      },
+    });
+  });
+
+  describe.each([
+    {
+      ruleType: 'query',
+      fieldName: 'rule_name_override',
+      humanizedFieldName: 'Rule name override',
+      initial: { field_name: 'fieldA' },
+      customized: { field_name: 'fieldB' },
+      upgrade: { field_name: 'fieldC' },
+      resolvedValue: { field_name: 'resolvedStringField' },
+    },
+  ] as const)(
+    '$fieldName ($ruleType rule)',
+    ({ ruleType, fieldName, humanizedFieldName, initial, customized, upgrade, resolvedValue }) => {
+      assertRuleUpgradePreview({
+        ruleType,
+        fieldName,
+        humanizedFieldName,
+        fieldVersions: {
+          initial,
+          customized,
+          upgrade,
+          resolvedValue,
+        },
+      });
+
+      assertRuleUpgradeAfterReview({
+        ruleType,
+        fieldName,
+        fieldVersions: {
+          initial,
+          customized,
+          upgrade,
+          resolvedValue,
+        },
+      });
+    }
+  );
+});
